Team console commands. Hydrator updates. App update

+++ public_html/migrations/20210130180703_create_teams_table.php
@@ -22,7 +22,7 @@ class CreateTeamsTable extends AbstractMigration
             $table->string('training_priority');
             $table->string('status');
             $table->double('money');
-            $table->double('rating');
+            $table->integer('rating');
             $table->integer('is_default')->defaultValue(1);
 
             $table->index('user_id');
